+++ src/handlers/frontend/context.rs
@@ -10,7 +10,7 @@ pub struct IndexContext {
     pub show_join_modal: bool,
     pub is_rejoin_page: bool,
     pub auth_token: Option<String>,
-    pub player_id: Option<i64>,
+    pub player_id: Option<i32>,
     pub player_name: Option<String>,
     pub rejoin_link: Option<String>,
 }
